# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  project.CreateProjectRequestBody:
    properties:
      description:
        type: string
      project_name:
        type: string
    required:
    - description
    - project_name
    type: object
...
  util.Res:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /project/create:
    post:
      consumes:
      - application/json
      description: Create Project
      operationId: CreateProject
      parameters:
      - description: Create Project
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: CreateProject
      tags:
      - CreateProject
  /project/delete:
    delete:
      consumes:
      - application/json
      description: Delete Project
      operationId: DeleteProject
      parameters:
      - description: Project_Name
        in: query
        name: project_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: DeleteProject
      tags:
      - DeleteProject
  /project/get_id:
    get:
      consumes:
      - application/json
      description: Get ProjectID
      operationId: GetProjectId
      parameters:
      - description: Project Name
        in: query
        name: project_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: GetProjectId
      tags:
      - GetProjectId
  /project/list:
    get:
      consumes:
      - application/json
      description: List Project
      operationId: ListProject
      parameters:
      - description: Project Name
        in: query
        name: project_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: ListProject
      tags:
      - ListProject
  /project/update:
    put:
      consumes:
      - application/json
      description: Update Project
      operationId: UpdateProject
      parameters:
      - description: Project_ID
        in: query
        name: project_id
        required: true
        type: string
      - description: Update Project
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      security:
      - ApiKeyAuth: []
      summary: UpdateProject
      tags:
      - UpdateProject
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"